kernel: security audit fixes — C1-C6, H3-H4, M5
Critical fixes:
- C1: procfs UAF — add g_pid_cmdline pool, stop overwriting g_pid_status
- C2: heap corruption — kill process + schedule() instead of infinite loop
- C3: ext2 consistency — write inode after each i_blocks increment
- C4: shm UAF — reject shm_at on IPC_RMID'd segments (-EIDRM),
  skip marked_rm segments in shm_get lookup
- C6: ELF W^X — parse p_flags for segment permissions, re-protect
  after relocations, only re-protect full pages within non-writable
  segments (partial pages may be shared with .data)

High fixes:
- H3: execve_copy_user_str — add upfront user_range_ok check
- H4: rq_remove_if_queued — scan all priority queues in both
  active and expired runqueues (not just current priority)

Low fixes:
- M5: frame_refcount uint16_t → uint32_t to prevent overflow

Also adds EIDRM/ENOMSG errno definitions.

Tests: 120/120 QEMU, 33/33 battery, 69/69 host — zero regressions

+++ src/mm/pmm.c
@@ -28,7 +28,7 @@ extern uint8_t _end;
 #define BITMAP_SIZE (MAX_RAM_SIZE / PAGE_SIZE / 8)
 
 static uint8_t memory_bitmap[BITMAP_SIZE];
-static uint16_t frame_refcount[MAX_RAM_SIZE / PAGE_SIZE];
+static uint32_t frame_refcount[MAX_RAM_SIZE / PAGE_SIZE];
 static uint64_t total_memory = 0;
 static uint64_t used_memory = 0;
 static uint64_t max_frames = 0;
@@ -210,7 +210,7 @@ void pmm_free_page(void* ptr) {
 
     uintptr_t flags = spin_lock_irqsave(&pmm_lock);
 
-    uint16_t rc = frame_refcount[frame];
+    uint32_t rc = frame_refcount[frame];
 
     if (rc == 0 || !bitmap_test(frame)) {
         spin_unlock_irqrestore(&pmm_lock, flags);
@@ -240,11 +240,11 @@ void pmm_incref(uintptr_t paddr) {
     spin_unlock_irqrestore(&pmm_lock, flags);
 }
 
-uint16_t pmm_decref(uintptr_t paddr) {
+uint32_t pmm_decref(uintptr_t paddr) {
     uint64_t frame = paddr / PAGE_SIZE;
     if (frame == 0 || frame >= max_frames) return 0;
     uintptr_t flags = spin_lock_irqsave(&pmm_lock);
-    uint16_t new_val = --frame_refcount[frame];
+    uint32_t new_val = --frame_refcount[frame];
     if (new_val == 0) {
         bitmap_unset(frame);
         used_memory -= PAGE_SIZE;
@@ -253,11 +253,11 @@ uint16_t pmm_decref(uintptr_t paddr) {
     return new_val;
 }
 
-uint16_t pmm_get_refcount(uintptr_t paddr) {
+uint32_t pmm_get_refcount(uintptr_t paddr) {
     uint64_t frame = paddr / PAGE_SIZE;
     if (frame >= max_frames) return 0;
     uintptr_t flags = spin_lock_irqsave(&pmm_lock);
-    uint16_t rc = frame_refcount[frame];
+    uint32_t rc = frame_refcount[frame];
     spin_unlock_irqrestore(&pmm_lock, flags);
     return rc;
 }
